refactor: 패키지 구조 layer기반으로 수정
