+++ src/main/java/wooteco/subway/controller/StationController.java
@@ -1,0 +1,63 @@
+package wooteco.subway.controller;
+
+import java.net.URI;
+import java.sql.SQLException;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import wooteco.subway.dao.station.StationDao;
+import wooteco.subway.domain.station.Station;
+import wooteco.subway.controller.dto.request.StationRequest;
+import wooteco.subway.controller.dto.response.StationResponse;
+
+@RestController
+public class StationController {
+
+    @Autowired
+    private StationDao stationDao;
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<StationResponse> exceptionHandler(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(SQLException.class)
+    public ResponseEntity<StationResponse> sqlExceptionHandler(SQLException e) {
+        return ResponseEntity.badRequest().build();
+    }
+
+    @PostMapping("/stations")
+    public ResponseEntity<StationResponse> createStation(
+        @RequestBody StationRequest stationRequest) {
+        Station station = new Station(stationRequest.getName());
+        Station newStation = stationDao.save(station);
+        StationResponse stationResponse = new StationResponse(newStation.getId(),
+            newStation.getName());
+        return ResponseEntity.created(URI.create("/stations/" + newStation.getId()))
+            .body(stationResponse);
+    }
+
+    @GetMapping(value = "/stations", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<StationResponse>> showStations() {
+        List<Station> stations = stationDao.findAll();
+        List<StationResponse> stationResponses = stations.stream()
+            .map(it -> new StationResponse(it.getId(), it.getName()))
+            .collect(Collectors.toList());
+        return ResponseEntity.ok().body(stationResponses);
+    }
+
+    @DeleteMapping("/stations/{id}")
+    public ResponseEntity deleteStation(@PathVariable Long id) {
+        stationDao.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+}
