+++ src/main/java/wooteco/subway/controller/dto/request/SectionRequest.java
@@ -1,0 +1,29 @@
+package wooteco.subway.controller.dto.request;
+
+public class SectionRequest {
+
+    private Long upStationId;
+    private Long downStationId;
+    private int distance;
+
+    public SectionRequest() {
+    }
+
+    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
