+++ src/main/java/wooteco/subway/controller/dto/request/StationRequest.java
@@ -1,0 +1,17 @@
+package wooteco.subway.controller.dto.request;
+
+public class StationRequest {
+
+    private String name;
+
+    public StationRequest() {
+    }
+
+    public StationRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
