+++ src/main/java/wooteco/subway/controller/dto/response/LineResponse.java
@@ -1,0 +1,34 @@
+package wooteco.subway.controller.dto.response;
+
+import java.util.List;
+
+public class LineResponse {
+
+    private final Long id;
+    private final String name;
+    private final String color;
+    private final List<StationResponse> stations;
+
+    public LineResponse(Long id, String name, String color, List<StationResponse> stations) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+        this.stations = stations;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
+    }
+}
