+++ src/main/java/wooteco/subway/controller/dto/response/StationResponse.java
@@ -1,0 +1,23 @@
+package wooteco.subway.controller.dto.response;
+
+public class StationResponse {
+
+    private Long id;
+    private String name;
+
+    public StationResponse() {
+    }
+
+    public StationResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
