--- src/main/java/wooteco/subway/line/LineResponse.java
@@ -1,35 +1,0 @@
-package wooteco.subway.line;
-
-import java.util.List;
-import wooteco.subway.station.StationResponse;
-
-public class LineResponse {
-
-    private final Long id;
-    private final String name;
-    private final String color;
-    private final List<StationResponse> stations;
-
-    public LineResponse(Long id, String name, String color, List<StationResponse> stations) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.stations = stations;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public List<StationResponse> getStations() {
-        return stations;
-    }
-}
