--- src/main/java/wooteco/subway/line/dao/LineDao.java
@@ -1,17 +1,0 @@
-package wooteco.subway.line.dao;
-
-import java.util.List;
-import wooteco.subway.line.Line;
-
-public interface LineDao {
-
-    Line save(Line line);
-
-    List<Line> findAll();
-
-    Line findById(Long id);
-
-    void update(Line updatedLine);
-
-    void deleteById(Long id);
-}
