--- src/main/java/wooteco/subway/station/dao/StationDao.java
@@ -1,15 +1,0 @@
-package wooteco.subway.station.dao;
-
-import java.util.List;
-import wooteco.subway.station.Station;
-
-public interface StationDao {
-
-    Station save(Station station);
-
-    List<Station> findAll();
-
-    void deleteById(Long id);
-
-    Station findById(Long id);
-}
